requirements.txt
flask
opencv-python
numpy
torch
torchvision
tensorflow
keras
Pillow
requests
